Move trial feature card out of unused component folder
