+++ client/my-sites/plans/ecommerce-trial/wx-medium-features.ts
@@ -5,7 +5,7 @@ import growth from 'calypso/assets/images/plans/wpcom/ecommerce-trial/growth.png
 import payments from 'calypso/assets/images/plans/wpcom/ecommerce-trial/payments.png';
 import productManagement from 'calypso/assets/images/plans/wpcom/ecommerce-trial/product-management.png';
 import shipping from 'calypso/assets/images/plans/wpcom/ecommerce-trial/shipping.png';
-import type { WooExpressMediumPlanFeatureSet } from 'calypso/my-sites/plans/components/ecommerce-plan-features/trial-feature-card';
+import type { WooExpressMediumPlanFeatureSet } from 'calypso/my-sites/plans/components/trial-feature-card';
 
 type WooExpressMediumFeatureSetProps = {
 	translate: typeof i18nTranslate;
